Fix written image-format
The default behaviour is to name all files as `<basename>.jpg` on the
assumption that all images are downloaded with type `image/jpeg`. But
this is not necessarily true, images may also be downloaded as
`image/png` in which case the file-name is wrong.

To fix this, add a new choice `native` to the option `--images-format`
and have it detect the actual downloaded image type from the
`Content-Type` header. Make this new choice the (changed) default
choice. Keep the choices `jpg` and `png` but enforce conversion in both
cases if the downloaded image does not match the desired format.